Address review round 2 comment 5: extract shared routing module
Move the public routing configuration types (LogsEventNameMapping,
SpanEventNameMapping and their routing keys), shared routing primitives, and
both resolution paths (log-view resolvers and proto-span resolvers, previously
split between otlp_encoder.rs and client.rs) into a new payload_encoder/routing
module. Public API paths are preserved via re-exports from client.rs.

Copilot-Session: bbdcce58-374c-4e6b-9f86-bc6ba6aaa5c1
