+++ opentelemetry-exporter-geneva/geneva-uploader/src/payload_encoder/mod.rs
@@ -2,6 +2,7 @@ pub(crate) mod bond_encoder;
 pub(crate) mod central_blob;
 pub(crate) mod lz4_chunked_compression;
 pub mod otlp_encoder;
+pub(crate) mod routing;
 
 #[cfg(test)]
 mod tests {
